ci: create Sentry release on publish (#7362)
Explicitly creating the Sentry release allows us to associate the
commits since the last release with the new one. This might help us to
identify potential sources of regressions. For the current releases,
I've set them manually to ensure that this automation has something to
pick up on for the next release.

The releases will already exists prior to this because they are
automatically created when a client / gateway first logs in with a
certain version.

What this does it mark it as "finalized" and set the commit range
accordingly.

Resolves: #7358.

---------

Signed-off-by: Thomas Eizinger <thomas@eizinger.io>
Co-authored-by: Jamil <jamilbk@users.noreply.github.com>