chore(snownet): print attributes for all requests and responses (#7380)
When debugging issues related to our TURN allocation code, we sometimes
only have the logs that code submitted to Sentry. As part of the event,
we submit the last 500 debug logs as breadcrumbs to give more context to
the error.

Unconditionally printing the attributes of each request-response pair
will help us in more easily diagnosing, why certain errors happen.