ci: set branch name as docker tag for images (#7389)
Our `docker-compose` file references the images using the `main` tag.
However, doing a `docker compose pull` fails because the `main` tag
doesn't exist for the client, gateway and relay images.

In order for this tag to exist, we need to instruct the
`docker/metadata-action` to generate a tag for the current branch.