chore(rust): add comments for ignored advisories (#7408)
We ignore some advisories of unmaintained crates flagged by `cargo
deny`. As long as the crates work for us, there is not much reason to
directly remove them, especially if it requires upstream effort. We will
get rid of these as they cause problems.

To avoid having to look up what they correspond to, we add a comment to
each line.